Initial support for library version.
